WEB PROGRAMMING
An online art community wants to create and manage a gallery of images they use for inspiration. They understand that this could be a large and complex site, but to begin with they want something simple to test if it interests their members.

I have been approached to create this initial version of the website for them.

The idea is that anybody should be able to view the image gallery.
Only community members with passwords should be able to log in and manage the gallery.
Managing the gallery means adding and deleting entries.
To begin with, the community is interested in allowing members to add the URL of an image and a brief description of it.

The community is full of people who care deeply about colour, layout, texture and space. However, they are completely unable to agree on a look-and-feel for the website, or even a logo.